FastElliptic API

Index

FastElliptic.E

FastElliptic.F

FastElliptic.J

FastElliptic.K

FastElliptic.Pi

FastElliptic.acn

FastElliptic.asn

FastElliptic.cn

FastElliptic.dn

FastElliptic.sn

Elliptic Integrals

FastElliptic.K — Function

K(m) -> Any

$K(m) = \int_0^{\pi/2}\frac{d\theta}{\sqrt{1-k^2\sin(\theta)^2}}.$

Returns the complete elliptic integral of the first kind.

Arguments

m : Elliptic modulus

FastElliptic.E — Function

E(m) -> Any

$E(m) = \int_0^{\pi/2}\sqrt{1-k^2\sin(\theta)^2}d\theta.$

Returns the complete elliptic integral of the second kind.

Arguments

m : Elliptic modulus

E(φ, m) -> Any

$E(\varphi, m) = \int_0^{\varphi}d\theta \sqrt{1-k^2\sin(\theta)^2}.$

Returns the incomplete elliptic integral of the second kind.

Arguments

φ : Amplitude

m : Elliptic modulus

FastElliptic.F — Function

F(φ, m) -> Any

$F(\varphi, m) = \int_0^{\varphi}\frac{d\theta}{\sqrt{1-k^2\sin(\theta)^2}}.$

Returns the incomplete elliptic integral of the first kind.

Arguments

φ : Amplitude

m : Elliptic modulus

FastElliptic.Pi — Function

Pi(n, m) -> Any

$\Pi(n;\varphi \,|\,m)=\int_{0}^{1 }{\frac{1}{1-nt^{2}}}{\frac{dt}{\sqrt{\left(1-mt^{2}\right)\left(1-t^{2}\right)}}}.$

Returns the complete elliptic integral of the third kind.

Arguments

n : Characteristic

m : Elliptic modulus

Pi(n, φ, m) -> Any

$\Pi (n;\varphi \,|\,m)=\int _{0}^{\sin \varphi }{\frac {1}{1-nt^{2}}}{\frac {dt}{\sqrt {\left(1-mt^{2}\right)\left(1-t^{2}\right)}}}.$

Returns the incomplete elliptic integral of the third kind.

Arguments

n : Characteristic

φ : Amplitude

m : Elliptic modulus

FastElliptic.J — Function

J(n, φ, m) -> Any

$J (n;\varphi \,|\,m)=\frac{\Pi(n;\varphi|\, m) - F(\varphi|\,m)}{n}.$

Returns the associate incomplete elliptic integral of the third kind.

Arguments

n : Characteristic

φ : Amplitude

m : Elliptic modulus

J(n, m) -> Any

$J(n;\varphi \,|\,m)=\frac{\Pi(n;\pi/2|\, m) - K(m)}{n}.$

Returns the associate complete elliptic integral of the third kind.

Arguments

n : Characteristic

m : Elliptic modulus

Jacobi Elliptic Functions

FastElliptic.sn — Function

sn(u, m) -> Any

$\text{sn}(u,m)=\sin\,\text{am}(u,m)$, where $\text{am}(u|\,m)=F^{-1}(u|\,m)$ is the Jacobi amplitude.

Returns the Jacobi Elliptic sn.

Arguments

u : Amplitude

m : Elliptic modulus

FastElliptic.cn — Function

cn(u, m) -> Any

$\text{cn}(u,m)=\cos\,\text{am}(u,m)$, where $\text{am}(u|\,m)=F^{-1}(u|\,m)$ is the Jacobi amplitude.

Returns the Jacobi Elliptic cn.

Arguments

u : Amplitude

m : Elliptic modulus

FastElliptic.dn — Function

dn(u, m) -> Any

$\text{dn}(u,m)=\frac{d}{d u}\,\text{am}(u,m)$, where $\text{am}(u|\,m)=F^{-1}(u|\,m)$ is the Jacobi amplitude.

Returns the Jacobi Elliptic dn.

Arguments

u : Amplitude

m : Elliptic modulus

FastElliptic.asn — Function

asn(s, m) -> Any

$\text{asn}(u,m)=\text{sn}(u,m)^{-1}$.

Returns the inverse Jacobi Elliptic sn.

Arguments

u : Amplitude

m : Elliptic modulus

FastElliptic.acn — Function

acn(c, m) -> Any

$\text{acn}(u,m)=\text{cn}(u,m)^{-1}$.

Returns the inverse Jacobi Elliptic cn.

Arguments

u : Amplitude

m : Elliptic modulus