JacobiElliptic API

Index

JacobiElliptic.CarlsonAlg.am

JacobiElliptic.CarlsonAlg.ellipke

JacobiElliptic.FukushimaAlg.E

JacobiElliptic.FukushimaAlg.F

JacobiElliptic.FukushimaAlg.J

JacobiElliptic.FukushimaAlg.K

JacobiElliptic.FukushimaAlg.Pi

JacobiElliptic.FukushimaAlg.acn

JacobiElliptic.FukushimaAlg.asn

JacobiElliptic.FukushimaAlg.cn

JacobiElliptic.FukushimaAlg.dn

JacobiElliptic.FukushimaAlg.sn

Elliptic Integrals

JacobiElliptic.FukushimaAlg.K — Function

$K(m) = \int_0^{\pi/2}\frac{d\theta}{\sqrt{1-k^2\sin(\theta)^2}}.$

Returns the complete elliptic integral of the first kind.

Arguments

m : Elliptic modulus

JacobiElliptic.FukushimaAlg.E — Function

$E(m) = \int_0^{\pi/2}\sqrt{1-k^2\sin(\theta)^2}d\theta.$

Returns the complete elliptic integral of the second kind.

Arguments

m : Elliptic modulus

$E(\varphi|\, m) = \int_0^{\varphi}d\theta \sqrt{1-m\sin(\theta)^2}.$

Returns the incomplete elliptic integral of the second kind.

Arguments

φ : Amplitude

m : Elliptic modulus

JacobiElliptic.FukushimaAlg.F — Function

$F(\varphi |\, m) = \int_0^{\varphi}\frac{d\theta}{\sqrt{1-m\sin(\theta)^2}}.$

Returns the incomplete elliptic integral of the first kind.

Arguments

φ : Amplitude

m : Elliptic modulus

JacobiElliptic.FukushimaAlg.Pi — Function

$\Pi(n|\,m)=\int_{0}^{1 }{\frac{1}{1-nt^{2}}}{\frac{dt}{\sqrt{\left(1-mt^{2}\right)\left(1-t^{2}\right)}}}.$

Returns the complete elliptic integral of the third kind.

Arguments

n : Characteristic

m : Elliptic modulus

$\Pi (n;\varphi \,|\,m)=\int _{0}^{\sin \varphi }{\frac {1}{1-nt^{2}}}{\frac {dt}{\sqrt {\left(1-mt^{2}\right)\left(1-t^{2}\right)}}}.$

Returns the incomplete elliptic integral of the third kind.

Arguments

n : Characteristic

φ : Amplitude

m : Elliptic modulus

JacobiElliptic.FukushimaAlg.J — Function

$J (n;\varphi \,|\,m)=\frac{\Pi(n;\varphi|\, m) - F(\varphi|\,m)}{n}.$

Returns the associate incomplete elliptic integral of the third kind.

Arguments

n : Characteristic

φ : Amplitude

m : Elliptic modulus

$J(n;\varphi \,|\,m)=\frac{\Pi(n;\pi/2|\, m) - K(m)}{n}.$

Returns the associate complete elliptic integral of the third kind.

Arguments

n : Characteristic

m : Elliptic modulus

Jacobi Elliptic Functions

JacobiElliptic.FukushimaAlg.sn — Function

$\text{sn}(u|\,m)=\sin\,\text{am}(u,m)$, where $\text{am}(u|\,m)=F^{-1}(u|\,m)$ is the Jacobi amplitude.

Returns the Jacobi Elliptic sn.

Arguments

u : Amplitude

m : Elliptic modulus

JacobiElliptic.FukushimaAlg.cn — Function

$\text{cn}(u|\,m)=\cos\,\text{am}(u,m)$, where $\text{am}(u|\,m)=F^{-1}(u|\,m)$ is the Jacobi amplitude.

Returns the Jacobi Elliptic cn.

Arguments

u : Amplitude

m : Elliptic modulus

JacobiElliptic.FukushimaAlg.dn — Function

$\text{dn}(u|\,m)=\frac{d}{d u}\,\text{am}(u,m)$, where $\text{am}(u|\,m)=F^{-1}(u|\,m)$ is the Jacobi amplitude.

Returns the Jacobi Elliptic dn.

Arguments

u : Amplitude

m : Elliptic modulus

JacobiElliptic.FukushimaAlg.asn — Function

$\text{asn}(u|\,m)=\text{sn}^{-1}(u,m)$.

Returns the inverse Jacobi Elliptic sn.

Arguments

u : Amplitude

m : Elliptic modulus

JacobiElliptic.FukushimaAlg.acn — Function

$\text{acn}(u|\,m)=\text{cn}^{-1}(u,m)$.

Returns the inverse Jacobi Elliptic cn.

Arguments

u : Amplitude

m : Elliptic modulus

JacobiElliptic.CarlsonAlg.am — Function

am(u::Real, m::Real, [tol::Real=eps(Float64)])

Returns amplitude, φ, such that u = F(φ | m)

Landen sequence with convergence to tol used if √(tol) ≤ m ≤ 1 - √(tol)

JacobiElliptic.CarlsonAlg.ellipke — Function

ellipke(m::Real) returns (K(m), E(m)) for scalar 0 ≤ m ≤ 1